+++ UnityAds/UnityAdsInterstitialCustomEvent.m
@@ -11,6 +11,7 @@
 #if __has_include("MoPub.h")
     #import "MPLogging.h"
 #endif
+#import "UnityAdsAdapterConfiguration.h"
 
 static NSString *const kMPUnityInterstitialVideoGameId = @"gameId";
 static NSString *const kUnityAdsOptionPlacementIdKey = @"placementId";
@@ -38,10 +39,18 @@ static NSString *const kUnityAdsOptionZoneIdKey = @"zoneId";
         self.placementId = [info objectForKey:kUnityAdsOptionZoneIdKey];
     }
     if (gameId == nil || self.placementId == nil) {
-        [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:[NSError errorWithDomain:@"" code:-1200 userInfo:@{NSLocalizedDescriptionKey: @"Custom event class data did not contain gameId/placementId.", NSLocalizedRecoverySuggestionErrorKey: @"Update your MoPub custom event class data to contain a valid Unity Ads gameId/placementId."}]];
+        NSError *error = [NSError errorWithCode:MOPUBErrorAdapterFailedToLoadAd localizedDescription:@"Unity Ads adapter is configured with an invalid placement id"];
+        MPLogAdEvent([MPLogEvent adLoadFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementId);
+        [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:error];
+
         return;
     }
+    
+    // Only need to cache game ID for SDK initialization
+    [UnityAdsAdapterConfiguration updateInitializationParameters:info];
+
     [[MPUnityRouter sharedRouter] requestVideoAdWithGameId:gameId placementId:self.placementId delegate:self];
+    MPLogAdEvent([MPLogEvent adLoadAttemptForAdapter:NSStringFromClass(self.class) dspCreativeId:nil dspName:nil], self.placementId);
 }
 
 - (BOOL)hasAdAvailable
@@ -52,10 +61,13 @@ static NSString *const kUnityAdsOptionZoneIdKey = @"zoneId";
 - (void)showInterstitialFromRootViewController:(UIViewController *)viewController
 {
     if ([self hasAdAvailable]) {
+        MPLogAdEvent([MPLogEvent adShowAttemptForAdapter:NSStringFromClass(self.class)], self.placementId);
         [[MPUnityRouter sharedRouter] presentVideoAdFromViewController:viewController customerId:nil placementId:self.placementId settings:nil delegate:self];
     } else {
-        MPLogInfo(@"Failed to show Unity Interstitial: Unity now claims that there is no available video ad.");
-        [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:nil];
+        NSError *error = [NSError errorWithCode:MOPUBErrorAdapterFailedToLoadAd localizedDescription:@"Failed to show Unity Interstitial: Unity now claims that there is no available video ad."];
+        
+        MPLogAdEvent([MPLogEvent adShowFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementId);
+        [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:error];
     }
 }
 
@@ -79,35 +91,47 @@ static NSString *const kUnityAdsOptionZoneIdKey = @"zoneId";
 {
     if (self.loadRequested) {
         [self.delegate interstitialCustomEvent:self didLoadAd:placementId];
+        MPLogAdEvent([MPLogEvent adLoadSuccessForAdapter:NSStringFromClass(self.class)], self.placementId);
         self.loadRequested = NO;
     }
 }
 
 - (void)unityAdsDidError:(UnityAdsError)error withMessage:(NSString *)message
 {
-    [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:nil];
+    NSError *loadError = [NSError errorWithCode:MOPUBErrorAdapterFailedToLoadAd localizedDescription:@"Unity Ads failed to load an ad"];
+    [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:loadError];
+    MPLogAdEvent([MPLogEvent adLoadFailedForAdapter:NSStringFromClass(self.class) error:loadError], self.placementId);
 }
 
 - (void) unityAdsDidStart:(NSString *)placementId
 {
     [self.delegate interstitialCustomEventWillAppear:self];
+    MPLogAdEvent([MPLogEvent adWillAppearForAdapter:NSStringFromClass(self.class)], self.placementId);
+    MPLogAdEvent([MPLogEvent adShowSuccessForAdapter:NSStringFromClass(self.class)], self.placementId);
+
     [self.delegate interstitialCustomEventDidAppear:self];
+    MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], self.placementId);
 }
 
 - (void) unityAdsDidFinish:(NSString *)placementId withFinishState:(UnityAdsFinishState)state
 {
     [self.delegate interstitialCustomEventWillDisappear:self];
+    MPLogAdEvent([MPLogEvent adWillDisappearForAdapter:NSStringFromClass(self.class)], self.placementId);
+
     [self.delegate interstitialCustomEventDidDisappear:self];
+    MPLogAdEvent([MPLogEvent adDidDisappearForAdapter:NSStringFromClass(self.class)], self.placementId);
 }
 
 - (void) unityAdsDidClick:(NSString *)placementId
 {
     [self.delegate interstitialCustomEventDidReceiveTapEvent:self];
+    MPLogAdEvent([MPLogEvent adTappedForAdapter:NSStringFromClass(self.class)], self.placementId);
 }
 
 - (void)unityAdsDidFailWithError:(NSError *)error
 {
     [self.delegate interstitialCustomEvent:self didFailToLoadAdWithError:error];
+    MPLogAdEvent([MPLogEvent adLoadFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementId);
 }
 
 @end